validator_creditcard.cpp
In this digital world where e-commerce is ruling, credit cards are a must-have. Most payment gateways utilize some kind of validation mechanism for credit card data before they proceed with the payment. What if there exists a software to accomplish this validation from the get-go? The validation operates differently for different cards, so deciding what type of card it is and then inspecting if the number entered is a valid credit card number or not is the way to go.If things turn to be a bit too tricky from the get-go, it’s okay to enforce a drop-down or menu for choosing which kind of card to check against. This project is quite simplistic in terms of design but will need some knowledge of algorithms if one proceeds with creating it.