banking system
The main objective of this banking system which is developed in c++ is to provide a flexible solutions for banking industry. With this system both type of users such as bank customers and working personnel of the particular bank will find easy to use and able to perform all their operations using this system. This system enable its users to open and close the bank account. Customers will able to withdraw and deposit amount to their particular account. It will also able to generate transactions on money for particular customer account.